Add check for sudo

+++ home/pi/scripts/packitupandgo.sh
@@ -1,5 +1,11 @@
 #!/bin/bash
 
+if [ $(id -u) -ne 0 ]; then
+    printf "\n"
+    printf "script must be run as root\n"
+    exit 1
+fi
+
 rm -f /etc/wpa_supplicant/wpa_supplicant.conf;
 cp /home/pi/scripts/wpa_supplicant.conf /etc/wpa_supplicant/wpa_supplicant.conf;
 sudo find /var/log -type f -delete;
